requirements.txt
jsonlines
openai
langchain
langchain-openai
langchain-community
beautifulsoup4
langchain_chroma
chromadb==0.5.3
pymupdf
sentence-transformers
faiss-cpu

# .env 설정
python-dotenv